+++ modules/graphics/src/shader.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "vscge/core/shader.h"
+#include "vscge/graphics/shader.h"
 
 #include <gl/gl3w.h>