+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.17.0] - 2020-09-03
+
+### Added
+
+- `v` shorthands for `Vec` classes.
+- `graphics` module.
+- `VertexBuffer` class.
+- `VSCGEModules.cmake` with useful `macro`s for CMake.
+
+### Changed
+
+- `ui` shorthands to just `u`.
+- Moved `Shader`, `Renderer`, and `Vertex` to `graphics`.
+- Revamped non-interface modules' `CMakeLists.txt` to use `GLOB`, sue me.
+
 ## [0.16.1] - 2020-08-31
 
 ### Added